Move some component
