+++ src/stories/containers/RecognizedDelegates/TotalAndKeyStatsComponent/Range.tsx
@@ -1,7 +1,7 @@
 import styled from '@emotion/styled';
 
 import React from 'react';
-import ChipRange from './ChipRange';
+import ChipRange from '../ChipRange';
 
 interface Props {
   start: string;
